fastapi_mysql/pyproject.toml
[project]
name = "fastapi-mysql"
version = "0.1.0"
description = ""
authors = [
    {name = "sigirace"}
]
readme = "README.md"
requires-python = ">=3.11, < 4.0"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "debugpy (>=1.8.14,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "starlette-context (>=0.4.0,<0.5.0)",
    "dependency-injector (>=4.47.1,<5.0.0)",
    "asyncmy (>=0.2.10,<0.3.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "bcrypt (==3.2.2)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pymysql (>=1.1.1,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "pytest-mock (>=3.14.1,<4.0.0)",
    "freezegun (>=1.5.2,<2.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
